+++ sdk/identity/Azure.Identity.Broker/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Support Microsoft Broker on macOS.
+
 ### Breaking Changes
 
 ### Bugs Fixed